Simplifying all

+++ scripts/build.sh
@@ -1,67 +1,78 @@
 #/bin/bash
 
-
-# to test it
+# Example:
 # PYTHON_VERSION=3.7 scripts/build.sh
 
-# very verbose
-set -x
-
-# stop on error
+ # stop or error
 set -e
 
-# stop of the first command that fails with pipe (|)
-set -o pipefail
+# very verbose
+set -x 
 
-PYTHON_VERSION=${PYTHON_VERSION:-3.6}
+if [[ ${PYTHON_VERSION} == "" ]] ; then
+	echo "Please specify PYTHON_VERSION"
+	exit 1
+fi
 
 if [ "$(uname)" == "Darwin" ]; then 
 
-	if [ !  -x "$(command -v brew)" ]; then
-		/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+	CMAKE_OSX_SYSROOT=~/MacOSX-SDKs/MacOSX10.9.sdk
+	Qt5_DIR=/usr/local/opt/qt
+	PYTHON_EXECUTABLE=/usr/local/opt/python@${PYTHON_VERSION}/bin/python${PYTHON_VERSION}
+	
+	if [[ ! -x "$(command -v cmake)" ]] ; then
+		brew install cmake
 	fi
 	
-	brew update 1>/dev/null 2>/dev/null || true
-	brew install cmake swig || true
+	if [[ ! -x "$(command -v swig)" ]] ; then
+		brew install swig
+	fi	
 	
-	# install python
-	brew install sashkab/python/python@${PYTHON_VERSION} 
-	PYTHON_EXECUTABLE=$(brew --prefix python@${PYTHON_VERSION})/bin/python${PYTHON_VERSION}
-	alias python=${PYTHON_EXECUTABLE}
+	if [ ! -f "${PYTHON_EXECUTABLE}" ]; then
+		brew install sashkab/python/python@${PYTHON_VERSION} 
+		${PYTHON_EXECUTABLE} -m pip install -q --upgrade pip  || true
+		${PYTHON_EXECUTABLE} -m pip install -q numpy setuptools wheel twine	 || true
+	fi
 	
-	# some package I need
-	python -m pip install -q --upgrade pip  || true
-	python -m pip install -q numpy setuptools wheel twine	 || true
+	if [ ! -d "${Qt5_DIR}" ] ; then
+		brew install "https://raw.githubusercontent.com/Homebrew/homebrew-core/5eb54ced793999e3dd3bce7c64c34e7ffe65ddfd/Formula/qt.rb"  
+	fi
 	
-	# qt5 5.11.2
-	brew install "https://raw.githubusercontent.com/Homebrew/homebrew-core/5eb54ced793999e3dd3bce7c64c34e7ffe65ddfd/Formula/qt.rb"  
-	
-	CMAKE_OSX_SYSROOT=~/MacOSX-SDKs/MacOSX10.9.sdk	
-	if [ ! -d ${CMAKE_OSX_SYSROOT} ] ; then
+	if [ ! -d "${CMAKE_OSX_SYSROOT}" ] ; then
 		pushd ~
 		git clone https://github.com/phracker/MacOSX-SDKs.git 
 		popd
 	fi
 	
-	mkdir -p build_travis
-	cd build_travis
-	cmake -GXcode -DPYTHON_EXECUTABLE=${PYTHON_EXECUTABLE} -DCMAKE_OSX_SYSROOT=${CMAKE_OSX_SYSROOT} -DQt5_DIR=$(brew --prefix Qt)/lib/cmake/Qt5 ../
-	
-	if [[ "${TRAVIS_OS_NAME}" == "osx" ]] ; then
-		cmake --build ./ --target ALL_BUILD --config Release | xcpretty -c
-	else
-		cmake --build ./ --target ALL_BUILD --config Release
+	if [[ ! -d ~/miniconda3 ]] ; then
+		pushd ${HOME}
+		curl -L --insecure --retry 3 "https://repo.continuum.io/miniconda/Miniconda3-latest-MacOSX-x86_64.sh" -O		
+		bash ./Miniconda3-latest-MacOSX-x86_64.sh -b # b stands for silent mode
+		rm ./Miniconda3-latest-MacOSX-x86_64.sh
+		popd
+		
+		~/miniconda3/bin/conda config  --set changeps1 no --set anaconda_upload no
+		~/miniconda3/bin/conda update  --yes conda
+		~/miniconda3/bin/conda install --yes anaconda-client			
+		
 	fi
 	
-	cmake --build . --target install   --config Release
+	# install python inside conda
+	./scripts/install/python.conda.sh ${PYTHON_VERSION}
+	
+	
+	
+	mkdir -p build_travis
+	cd build_travis
+	cmake -GXcode -DPYTHON_EXECUTABLE=${PYTHON_EXECUTABLE} -DCMAKE_OSX_SYSROOT=${CMAKE_OSX_SYSROOT} -DQt5_DIR=${Qt5_DIR} ../
+	cmake --build ./ --target ALL_BUILD --config Release
+	cmake --build ./ --target install   --config Release
 
 else
 
-	PYTHON_EXECUTABLE=python${PYTHON_VERSION}
-	alias python=PYTHON_EXECUTABLE
-	
 	mkdir -p build_travis
 	cd build_travis
+	PYTHON_EXECUTABLE=python${PYTHON_VERSION}
 	cmake -DPYTHON_EXECUTABLE=${PYTHON_EXECUTABLE} -DQt5_DIR=/opt/qt59 ../
 	cmake --build ./ --target all     --config Release
 	cmake --build ./ --target install --config Release
@@ -73,7 +84,7 @@ cd Release/OpenVisus
 # test
 Tests="Array.py Dataflow.py Dataflow2.py Idx.py XIdx.py DataConversion1.py DataConversion2.py" 
 for Test in ${Tests} ; do
-	PYTHONPATH=../ python Samples/python/${Test}
+	PYTHONPATH=../ ${PYTHON_EXECUTABLE} Samples/python/${Test}
 done
 
 if [ "$(uname)" == "Darwin" ]; then
@@ -88,7 +99,7 @@ fi
 BUILD_WHEEL=${BUILD_WHEEL:-1}
 if (( BUILD_WHEEL == 1 )) ; then
 
-	python -m pip install setuptools wheel --upgrade || true
+	${PYTHON_EXECUTABLE} -m pip install setuptools wheel --upgrade || true
 	
 	if [ "$(uname)" == "Darwin" ]; then 
 		PLATFORM_TAG=macosx_10_9_x86_64
@@ -97,25 +108,19 @@ if (( BUILD_WHEEL == 1 )) ; then
 	fi	
 	
 	rm -Rf ./dist/*OpenVisus-*.whl 
-	python setup.py -q bdist_wheel --python-tag=cp${PYTHON_VERSION:0:1}${PYTHON_VERSION:2:1} --plat-name=${PLATFORM_TAG}
+	${PYTHON_EXECUTABLE} setup.py -q bdist_wheel --python-tag=cp${PYTHON_VERSION:0:1}${PYTHON_VERSION:2:1} --plat-name=${PLATFORM_TAG}
 	WHEEL_FILENAME=dist/OpenVisus-*.whl
 	rm -Rf ./build
 		
 	if [[ "${PYPI_DEPLOY}" != ""  && "${PYPI_USERNAME}" != "" && "${PYPI_PASSWORD}" != "" ]] ; then
-		python -m twine upload --username ${PYPI_USERNAME} --password ${PYPI_PASSWORD} --skip-existing ${WHEEL_FILENAME}
+		${PYTHON_EXECUTABLE} -m twine upload --username ${PYPI_USERNAME} --password ${PYPI_PASSWORD} --skip-existing ${WHEEL_FILENAME}
 	fi
 fi
 
-unalias python
-
 # ////////////////////////////////////////////////////////////////////////
 BUILD_CONDA=${BUILD_CONDA:-1}
-if (( BUILD_CONDA == 1 )) ; then
 
-	if [ "$(uname)" == "Darwin" ]; then 
-		./scripts/install/miniconda_osx.sh
-		./scripts/install/python.conda.sh ${PYTHON_VERSION}
-	fi
+if (( BUILD_CONDA == 1 )) ; then
 
 	# activate 
 	export PATH=~/miniconda3/bin:$PATH
@@ -134,14 +139,11 @@ if (( BUILD_CONDA == 1 )) ; then
 	CONDA_FILENAME=$(find ~/miniconda3/conda-bld -iname "openvisus*.tar.bz2")
 	rm -Rf ./build
 	
-	if (( 1 == 1 )) ; then
-		conda install -y ${CONDA_FILENAME}
-		
-		for Test in ${Tests} ; do
-			python $(python -m OpenVisus dirname)/Samples/python/${Test}
-		done
+	conda install -y ${CONDA_FILENAME}
 	
-	fi
+	for Test in ${Tests} ; do
+		python $(python -m OpenVisus dirname)/Samples/python/${Test}
+	done
 	
 	if [[ "${CONDA_DEPLOY}" != ""  && "${ANACONDA_TOKEN}" != "" ]] ; then
 		anaconda -t ${ANACONDA_TOKEN} upload "${CONDA_BUILD_FILENAME}"
